Refactor Docker Postgres files, add scripts to run
