+++ config.json
@@ -6,7 +6,7 @@
     "host": "localhost",
     "port": "5432",
     "user": "jieqiboh",
-    "pass": "",
+    "password": "postgres",
     "name": "patients",
     "sslmode": "disable"
   }
