+++ docker-compose.yaml
@@ -8,8 +8,10 @@ services:
     environment:
       - POSTGRES_USER=jieqiboh
       - POSTGRES_PASSWORD=postgres
+      - POSTGRES_DB=patients
     ports:
       - '5432:5432'
     volumes:
       # Mount and run the init.sql file to load data
-      - ./sqlstuff/init.sql:/docker-entrypoint-initdb.d/init.sql
+      - ./sql/init.sql:/docker-entrypoint-initdb.d/init.sql
+    container_name: postgres-docker-db
