+++ sql/data.json
@@ -1,0 +1,37 @@
+{
+  "admin": {
+    "familygroup": "Sample Family",
+    "name": "John Doe",
+    "age": 30,
+    "gender": "M"
+  },
+  "doctorsconsultation": {
+    "healthy": true,
+    "consultationnotes": null,
+    "referralneeded": false
+  },
+  "heightandweight": {
+    "height": 175.5,
+    "weight": 70.2
+  },
+  "pastmedicalhistory": {
+    "tuberculosis": true,
+    "diabetes": false,
+    "hypertension": true,
+    "hyperlipidemia": false,
+    "chronicjointpains": true
+  },
+  "socialhistory": {
+    "pastsmokinghistory": false,
+    "numberofyears": null,
+    "currentsmokinghistory": true
+  },
+  "visualacuity": {
+    "leyevision": 20,
+    "reyevision": 18
+  },
+  "vitalstatistics": {
+    "temperature": 98.6,
+    "spO2": 98.0
+  }
+}
