--- sqlstuff/patient_clean.sql
@@ -1,7 +1,0 @@
-DROP TABLE IF EXISTS pastmedicalhistory;
-DROP TABLE IF EXISTS socialhistory;
-DROP TABLE IF EXISTS vitalstatistics;
-DROP TABLE IF EXISTS heightandweight;
-DROP TABLE IF EXISTS visualacuity;
-DROP TABLE IF EXISTS doctorsconsultation;
-DROP TABLE IF EXISTS admin;
