+++ start.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# Start docker containers in detached mode, running in the background
+docker compose up -d
